// ExercicioAngular/noticias-app/src/app/models/news.models.ts
export interface Root {
  status: string
  totalResults: number
  articles: Article[]
}

export interface Article {
  source: Source
  author?: string
  title: string
  description?: string
  url: string
  urlToImage?: string
  publishedAt: string
  content?: string
}

export interface Source {
  id?: string
  name: string
}